Tiding and fix bugs
Former-commit-id: 05cd8c9d5a69f63aaaadf770ad8c19604db9597a

+++ population_model/src/main/java/uk/ac/standrews/cs/digitising_scotland/population_model/distributions/AffairSpacingDistribution.java
@@ -30,7 +30,7 @@ import uk.ac.standrews.cs.digitising_scotland.population_model.organic.OrganicPa
  */
 public final class AffairSpacingDistribution extends NormalDistribution {
 
-    private static final int STANDARD_DEVIATION_FACTOR = 4;
+    private static final double STANDARD_DEVIATION_FACTOR = 4;
 
     /**
      * Sets up an affair spacing distribution for the given partnership.
